Create Assignment_07
a) Write a C++ program to control a ball using
arrow keys. Apply the concept of polymorphism. OR (b) Write a C++ program to
implement bouncing ball using sine wave form. Apply the concept of
polymorphism. OR (c) Write C++ program to draw man walking in the rain with an
umbrella. Apply the concept of polymorphism. OR (c) Write a C++ program to
implement the game of 8 puzzle. Apply the concept of polymorphism. OR (d) Write
a C++ program to implement the game Tic Tac Toe. Apply the concept of
polymorphism.